moved prepareIndividualResponseCaseFromParentCase() to CaseService
