+++ src/test/java/uk/gov/ons/census/casesvc/service/FulfilmentRequestServiceTest.java
@@ -1,17 +1,17 @@
 package uk.gov.ons.census.casesvc.service;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.anyString;
+import static org.mockito.Mockito.eq;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 import static uk.gov.ons.census.casesvc.model.entity.EventType.FULFILMENT_REQUESTED;
 import static uk.gov.ons.census.casesvc.testutil.DataUtils.getRandomCase;
 import static uk.gov.ons.census.casesvc.testutil.DataUtils.getTestResponseManagementEvent;
 
-import java.time.OffsetDateTime;
-import java.util.ArrayList;
 import java.util.UUID;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
@@ -19,17 +19,9 @@ import uk.gov.ons.census.casesvc.logging.EventLogger;
 import uk.gov.ons.census.casesvc.model.dto.FulfilmentRequestDTO;
 import uk.gov.ons.census.casesvc.model.dto.ResponseManagementEvent;
 import uk.gov.ons.census.casesvc.model.entity.Case;
-import uk.gov.ons.census.casesvc.model.entity.CaseState;
 
 @RunWith(MockitoJUnitRunner.class)
 public class FulfilmentRequestServiceTest {
-  private static final String HOUSEHOLD_RESPONSE_ADDRESS_TYPE = "HH";
-  private static final String HOUSEHOLD_INDIVIDUAL_RESPONSE_ADDRESS_TYPE = "HI";
-  private static final String HOUSEHOLD_INDIVIDUAL_RESPONSE_REQUEST_ENGLAND = "UACIT1";
-  private static final String HOUSEHOLD_INDIVIDUAL_RESPONSE_REQUEST_WALES_ENGLISH = "UACIT2";
-  private static final String HOUSEHOLD_INDIVIDUAL_RESPONSE_REQUEST_WALES_WELSH = "UACIT2W";
-  private static final String HOUSEHOLD_INDIVIDUAL_RESPONSE_REQUEST_NORTHERN_IRELAND = "UACIT4";
-
   @Mock private EventLogger eventLogger;
 
   @Mock private CaseService caseService;
@@ -64,44 +56,21 @@ public class FulfilmentRequestServiceTest {
   }
 
   @Test
-  public void testGoodIndividualResponseFulfilmentRequestForUACIT1() {
-    testIndividualResponseCode(HOUSEHOLD_INDIVIDUAL_RESPONSE_REQUEST_ENGLAND);
-  }
-
-  @Test
-  public void testGoodIndividualResponseFulfilmentRequestForUACIT2() {
-    testIndividualResponseCode(HOUSEHOLD_INDIVIDUAL_RESPONSE_REQUEST_WALES_ENGLISH);
-  }
-
-  @Test
-  public void testGoodIndividualResponseFulfilmentRequestForUACIT2W() {
-    testIndividualResponseCode(HOUSEHOLD_INDIVIDUAL_RESPONSE_REQUEST_WALES_WELSH);
-  }
-
-  @Test
-  public void testGoodIndividualResponseFulfilmentRequestForUACIT4() {
-    testIndividualResponseCode(HOUSEHOLD_INDIVIDUAL_RESPONSE_REQUEST_NORTHERN_IRELAND);
-  }
-
-  private void testIndividualResponseCode(String individualResponseCode) {
+  public void testGoodIndividualResponseFulfilmentRequest() {
     // Given
-    Case parentCase = getRandomCase();
-    parentCase.setUacQidLinks(new ArrayList<>());
-    parentCase.setEvents(new ArrayList<>());
-    parentCase.setCreatedDateTime(OffsetDateTime.now().minusDays(1));
-    parentCase.setState(null);
-    parentCase.setReceiptReceived(true);
-    parentCase.setRefusalReceived(true);
-    parentCase.setAddressType("HH");
-
     ResponseManagementEvent managementEvent = getTestResponseManagementEvent();
     FulfilmentRequestDTO expectedFulfilmentRequest =
         managementEvent.getPayload().getFulfilmentRequest();
-    expectedFulfilmentRequest.setCaseId(parentCase.getCaseId().toString());
-    expectedFulfilmentRequest.setFulfilmentCode(individualResponseCode);
+    expectedFulfilmentRequest.setFulfilmentCode("UACIT1");
+
+    Case expectedCase = getRandomCase();
+    expectedFulfilmentRequest.setCaseId(expectedCase.getCaseId().toString());
+    Case expectedChildCase = getRandomCase();
 
     when(caseService.getCaseByCaseId(UUID.fromString(expectedFulfilmentRequest.getCaseId())))
-        .thenReturn(parentCase);
+        .thenReturn(expectedCase);
+    when(caseService.prepareIndividualResponseCaseFromParentCase(expectedCase))
+        .thenReturn(expectedChildCase);
 
     // when
     underTest.processFulfilmentRequest(managementEvent);
@@ -109,60 +78,13 @@ public class FulfilmentRequestServiceTest {
     // then
     verify(eventLogger, times(1))
         .logCaseEvent(
-            eq(parentCase),
+            eq(expectedCase),
             eq(managementEvent.getEvent().getDateTime()),
             eq("Fulfilment Request Received"),
             eq(FULFILMENT_REQUESTED),
             eq(managementEvent.getEvent()),
             anyString());
 
-    ArgumentCaptor<Case> caseArgumentCaptor = ArgumentCaptor.forClass(Case.class);
-    verify(caseService).saveAndEmitCaseCreatedEvent(caseArgumentCaptor.capture());
-    Case actualChildCase = caseArgumentCaptor.getValue();
-    checkIndivdualFulfilmentRequestCase(parentCase, actualChildCase);
-    verify(caseService, times(1)).getUniqueCaseRef();
-  }
-
-  private void checkIndivdualFulfilmentRequestCase(Case parentCase, Case actualChildCase) {
-    assertThat(actualChildCase.getCaseRef()).isNotEqualTo(parentCase.getCaseRef());
-    assertThat(UUID.fromString(actualChildCase.getCaseId().toString()))
-        .isNotEqualTo(parentCase.getCaseId());
-    assertThat(actualChildCase.getUacQidLinks()).isNull();
-    assertThat(actualChildCase.getEvents()).isNull();
-    assertThat(actualChildCase.getCreatedDateTime())
-        .isBetween(OffsetDateTime.now().minusSeconds(10), OffsetDateTime.now());
-    assertThat(actualChildCase.getCollectionExerciseId())
-        .isEqualTo(parentCase.getCollectionExerciseId());
-    assertThat(actualChildCase.getActionPlanId()).isEqualTo(parentCase.getActionPlanId());
-    assertThat(actualChildCase.getState()).isEqualTo(CaseState.ACTIONABLE);
-    assertThat(actualChildCase.isReceiptReceived()).isFalse();
-    assertThat(actualChildCase.isRefusalReceived()).isFalse();
-    assertThat(actualChildCase.getArid()).isEqualTo(parentCase.getArid());
-    assertThat(actualChildCase.getEstabArid()).isEqualTo(parentCase.getEstabArid());
-    assertThat(actualChildCase.getUprn()).isEqualTo(parentCase.getUprn());
-    assertThat(actualChildCase.getAddressType()).isEqualTo(HOUSEHOLD_RESPONSE_ADDRESS_TYPE);
-    assertThat(actualChildCase.getCaseType()).isEqualTo(HOUSEHOLD_INDIVIDUAL_RESPONSE_ADDRESS_TYPE);
-    assertThat(actualChildCase.getEstabType()).isEqualTo(parentCase.getEstabType());
-    assertThat(actualChildCase.getAddressLevel()).isNull();
-    assertThat(actualChildCase.getAbpCode()).isEqualTo(parentCase.getAbpCode());
-    assertThat(actualChildCase.getOrganisationName()).isEqualTo(parentCase.getOrganisationName());
-    assertThat(actualChildCase.getAddressLine1()).isEqualTo(parentCase.getAddressLine1());
-    assertThat(actualChildCase.getAddressLine2()).isEqualTo(parentCase.getAddressLine2());
-    assertThat(actualChildCase.getAddressLine3()).isEqualTo(parentCase.getAddressLine3());
-    assertThat(actualChildCase.getTownName()).isEqualTo(parentCase.getTownName());
-    assertThat(actualChildCase.getPostcode()).isEqualTo(parentCase.getPostcode());
-    assertThat(actualChildCase.getLatitude()).isEqualTo(parentCase.getLatitude());
-    assertThat(actualChildCase.getLongitude()).isEqualTo(parentCase.getLongitude());
-    assertThat(actualChildCase.getOa()).isEqualTo(parentCase.getOa());
-    assertThat(actualChildCase.getLsoa()).isEqualTo(parentCase.getLsoa());
-    assertThat(actualChildCase.getMsoa()).isEqualTo(parentCase.getMsoa());
-    assertThat(actualChildCase.getLad()).isEqualTo(parentCase.getLad());
-    assertThat(actualChildCase.getRegion()).isEqualTo(parentCase.getRegion());
-    assertThat(actualChildCase.getHtcWillingness()).isNull();
-    assertThat(actualChildCase.getHtcDigital()).isNull();
-    assertThat(actualChildCase.getFieldCoordinatorId()).isNull();
-    assertThat(actualChildCase.getFieldOfficerId()).isNull();
-    assertThat(actualChildCase.getTreatmentCode()).isNull();
-    assertThat(actualChildCase.getCeExpectedCapacity()).isNull();
+    verify(caseService).saveAndEmitCaseCreatedEvent(expectedChildCase);
   }
 }
